requirements.txt
python-mpd2==3.0.1
rapidfuzz>=0.12.3
